3-Code/Server_env_sensors/app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests>=2.32.0
apscheduler==3.7.0
mysql-connector-python>=9.1.0
Werkzeug>=3.0.6
SQLAlchemy==2.0.35
paho-mqtt==2.1.0